Extract compare-versions to file
